// pop-web/src/app/core/modules/openapi/model/tumor-marker-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeasureSchema } from './measure-schema';


export interface TumorMarkerCreate { 
    /**
     * Indicates the case of the patient related to the tumor marker result
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the tumor marker was analyzed.
     */
    date: string;
    /**
     * The chemical or biological substance/agent that is analyzed.
     */
    analyte: string;
    massConcentration?: MeasureSchema | null;
    arbitraryConcentration?: MeasureSchema | null;
    substanceConcentration?: MeasureSchema | null;
    fraction?: MeasureSchema | null;
    multipleOfMedian?: MeasureSchema | null;
    classification?: string | null;
    /**
     * References to the neoplastic entities that are related or the focus of the tumor marker analysis.
     */
    relatedEntitiesIds?: Array<string>;
}